Bullets for another time
The bullets are a bit strange. I don't know what else to use outside of
Math.atan2(double y, double x) to calculate my angle... I really don't
want to have to code that by hand.

For now it looks alright, and in terms of functionality, it works.
If I make the bullets go fast, nobody is going to notice anyways.
And if I make them really slow, I hope the users will realize that they
might be special bombs or something.